# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Microblogging API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts/{id}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a single post with its comments
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post with comments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostWithComments"
        "404":
          description: Not found
    patch:
      operationId: updatePost
      tags: [posts]
      summary: Update a post (owner only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePostBody"
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post (owner only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
...
  /feed-sources:
    get:
      operationId: listFeedSources
      tags: [feeds]
      summary: List configured RSS / Atom sources (owner only)
      responses:
        "200":
          description: Sources
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedSourcesList"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
    post:
      operationId: createFeedSource
      tags: [feeds]
      summary: Create a new feed source (owner only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFeedSourceBody"
      responses:
        "201":
          description: Created source
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedSource"
        "400":
          description: Invalid request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /feed-sources/{id}:
    patch:
      operationId: updateFeedSource
      tags: [feeds]
      summary: Update a feed source (owner only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFeedSourceBody"
      responses:
        "200":
          description: Updated source
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedSource"
        "400":
          description: Invalid request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
    delete:
      operationId: deleteFeedSource
      tags: [feeds]
      summary: Delete a feed source (owner only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
  /feed-sources/{id}/refresh:
    post:
      operationId: refreshFeedSource
      tags: [feeds]
      summary: Queue this feed source for an immediate background refresh (owner only)
      description: |
        Schedules a refresh on the background fetch worker and returns immediately.
        Outcome of the actual fetch (imported / skipped counts, upstream errors)
        lands in the source row's `last_status` / `last_error` columns and is
        visible on the next list reload. Returns `status: "ok"` when the work
        was queued; the response's `fetched` / `imported` / `skipped` counters
        are zero because the fetch has not run yet.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Refresh result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedRefreshResult"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
  /feed-sources/{id}/approve-all:
    post:
      operationId: approveAllFromFeedSource
      tags: [feeds]
      summary: Bulk-approve every pending post from this source (owner only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Number of pending posts approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApproveAllFromFeedSourceResponse"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
  /feed-sources/refresh:
    post:
      operationId: refreshAllFeedSources
      tags: [feeds]
      summary: Refresh every enabled, due source (owner cookie or X-Cron-Secret)
      parameters:
        - in: query
          name: force
          schema:
            type: string
            enum: ["1", "true"]
          description: When set, ignore each source's cadence and pull every enabled source.
      responses:
        "200":
          description: Aggregate refresh result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedRefreshSummary"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
...
components:
  schemas:
...
    Category:
      type: object
      description: Owner-managed taxonomy entry. Slug is the canonical identifier in URLs.
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - slug
        - name
        - createdAt
        - updatedAt
...
    Post:
      type: object
      properties:
        id:
          type: integer
        authorId:
          type: string
        authorName:
          type: string
        authorImageUrl:
          type: string
          nullable: true
        content:
          type: string
        contentFormat:
          type: string
          enum: [plain, html]
        commentCount:
          type: integer
        sourceFeedId:
          type: integer
          nullable: true
          description: ID of the feed_sources row that imported this post (PESOS); null for owner-authored posts.
        sourceFeedName:
          type: string
          nullable: true
          description: Display name of the originating feed source (joined from feed_sources). Null for owner-authored posts.
        sourceCanonicalUrl:
          type: string
          nullable: true
          description: Permalink of the post on its origin site (PESOS attribution).
        categories:
          type: array
          description: Categories this post belongs to (owner-managed taxonomy).
          items:
            $ref: "#/components/schemas/Category"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - authorId
        - authorName
        - content
        - contentFormat
        - commentCount
        - categories
        - createdAt
...
    Comment:
      type: object
      properties:
        id:
          type: integer
        postId:
          type: integer
        authorId:
          type: string
        authorName:
          type: string
        authorImageUrl:
          type: string
          nullable: true
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - postId
        - authorId
        - authorName
        - content
        - createdAt
    PostWithComments:
      type: object
      properties:
        post:
          $ref: "#/components/schemas/Post"
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
      required:
        - post
        - comments
...
    UpdatePostBody:
      type: object
      properties:
        content:
          type: string
          maxLength: 40000
        contentFormat:
          type: string
          enum: [plain, html]
        categoryIds:
          type: array
          description: |
            Replaces the post's category set. Sending an empty array clears all
            categories; omitting the field leaves the existing set untouched.
            Every id must be a positive integer; any unknown or malformed id
            causes the request to fail with 400 and the post stays unchanged.
          items:
            type: integer
            minimum: 1
      required:
        - content
        - contentFormat
...
    FeedSource:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        feedUrl:
          type: string
        siteUrl:
          type: string
          nullable: true
        cadence:
          type: string
          enum: [daily, weekly, monthly]
        enabled:
          type: boolean
        lastFetchedAt:
          type: string
          format: date-time
          nullable: true
        nextFetchAt:
          type: string
          format: date-time
          nullable: true
          description: When this source will next be eligible for refresh (null = never fetched, treated as immediately due).
        lastStatus:
          type: string
          nullable: true
        lastError:
          type: string
          nullable: true
        itemsImported:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - feedUrl
        - cadence
        - enabled
        - itemsImported
        - createdAt
        - updatedAt
    FeedSourcesList:
      type: object
      properties:
        sources:
          type: array
          items:
            $ref: "#/components/schemas/FeedSource"
      required:
        - sources
...
    CreateFeedSourceBody:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
        feedUrl:
          type: string
          format: uri
          maxLength: 2048
        siteUrl:
          type: string
          format: uri
          maxLength: 2048
          nullable: true
        cadence:
          type: string
          enum: [daily, weekly, monthly]
          default: daily
        enabled:
          type: boolean
          default: true
      required:
        - name
        - feedUrl
    UpdateFeedSourceBody:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
        feedUrl:
          type: string
          format: uri
          maxLength: 2048
        siteUrl:
          type: string
          format: uri
          maxLength: 2048
          nullable: true
        cadence:
          type: string
          enum: [daily, weekly, monthly]
        enabled:
          type: boolean
    FeedRefreshResult:
      type: object
      properties:
        sourceId:
          type: integer
        fetched:
          type: integer
        imported:
          type: integer
        skipped:
          type: integer
        status:
          type: string
          enum: [ok, error]
        error:
          type: string
          nullable: true
        alreadyInProgress:
          type: boolean
          description: |
            True when the per-source refresh endpoint was called while a
            background fetch for this source was already running. The
            endpoint returns `status: "ok"` with all counters at 0 and
            does not start a duplicate fetch.
      required:
        - sourceId
        - fetched
        - imported
        - skipped
        - status
    ApproveAllFromFeedSourceResponse:
      type: object
      properties:
        sourceId:
          type: integer
        approved:
          type: integer
          description: Number of pending posts that were flipped to published.
      required:
        - sourceId
        - approved
    FeedRefreshSummary:
      type: object
      properties:
        ranAt:
          type: string
          format: date-time
        attempted:
          type: integer
        totalFetched:
          type: integer
        totalImported:
          type: integer
        results:
          type: array
          items:
            $ref: "#/components/schemas/FeedRefreshResult"
      required:
        - ranAt
        - attempted
        - totalFetched
        - totalImported
        - results
...